feat(runtime-c-api) Extract the `global` module.
